backend/pyproject.toml
[tool.poetry]
name = "stripes"
version = "0.1.0"
description = "ShowYourLocalStripes.com"
authors = ["Javier SG <javiersaezgallego@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10.8"
fastapi = "^0.85.0"
uvicorn = { version = "^0.18.3", extras = ["standard"] }
pydantic = {extras = ["dotenv"], version = "^1.10.2"}
yarl = "^1.8.1"
ujson = "^5.5.0"
SQLAlchemy = {version = "^1.4.41", extras = ["mypy", "asyncio"]}
asyncpg = {version = "^0.26.0", extras = ["sa"]}
aiofiles = "^22.1.0"
httptools = "^0.5.0"
loguru = "^0.6.0"
pillow = "^9.3.0"
typer = "^0.7.0"
sqlmodel = "^0.0.8"
ipykernel = "^6.17.1"
psycopg2-binary = "^2.9.5"
alembic = "^1.8.1"
requests = "^2.28.1"
anyio =  "^3.6.2"
pytest-asyncio = "^0.20.2"
aiosqlite = "^0.17.0"
aiohttp = "^3.8.4"
httpx = "^0.23.0"
pandas = "^2.0.1"
matplotlib = "^3.7.1"
forex-python = "^1.8"
isort = "^5.12.0"
slack-sdk = "^3.22.0"
flake8 = "5.0.4"

[tool.poetry.dev-dependencies]
pytest = "^7.1.3"
mypy = "^0.981"
isort = "^5.10.1"
yesqa = "^1.4.0"
pre-commit = "^2.20.0"
black = "^22.8.0"
autoflake = "^1.6.1"
SQLAlchemy = {version = "^1.4.41", extras = ["mypy"]}
pytest-cov = "^4.0.0"
anyio = "^3.6.1"
pytest-env = "^0.6.2"
httpx = "^0.23.0"
pandas = "^2.0.1"
matplotlib = "^3.7.1"

[tool.isort]
profile = "black"
multi_line_output = 3
src_paths = ["api",]

[tool.mypy]
strict = true
ignore_missing_imports = true
allow_subclassing_any = true
allow_untyped_calls = true
pretty = true
show_error_codes = true
implicit_reexport = true
allow_untyped_decorators = true
warn_unused_ignores = false
warn_return_any = false
namespace_packages = true
plugins = ["sqlalchemy.ext.mypy.plugin"]

[tool.pytest.ini_options]
filterwarnings = [
    "error",
    "ignore::DeprecationWarning",
    "ignore:.*unclosed.*:ResourceWarning",
]
testpaths = [
    "tests"
]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
